+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## 0.24.7
 
+### Bug Fix
+- Fixed a bug that prevented newly created sheets from being locked
+
 ## 0.24.6 (May 7, 2014)
 
 ### Bug Fix